requirements.txt
pypdf>=4.0.0
ragas>=0.1.7
openai>=1.10.0
python-dotenv>=1.0.0
pandas>=2.1.0
tqdm>=4.66.0
jinja2>=3.1.0
google-generativeai>=0.3.0
